:recycle: Refactored models
